fix npm install issues (#16)

+++ npm/install.js
@@ -23,12 +23,19 @@ if (!fs.existsSync(bin)){
     fs.mkdirSync(bin);
 }
 
+let retries = 0;
+const MAX_RETRIES = 3;
+
 const install = () => {
   const tmpdir = os.tmpdir();
   const req = request({ uri: tarUrl });
 
-  const tarFile = `${tmpdir}/${name}.tar.gz`;
+  const tarFile = `${tmpdir}/${name}-${Date.now()}.tar.gz`;
   const download = fs.createWriteStream(tarFile);
+
+  if (retries > 0) {
+    console.log(`retrying to install safebox - retry ${retries} out of ${MAX_RETRIES}`)
+  }
 
   console.log(`downloading safebox binary`);
 
@@ -41,8 +48,23 @@ const install = () => {
   });
 
   req.on('complete', () => {
-    cp.execSync(`tar -xf ${tarFile} -C ${tmpdir}`);
-    fs.copyFileSync(path.join(tmpdir, binaryName), path.join(bin, binaryName));
+    console.log('download complete. installing safebox.')
+
+    try {
+      if (!fs.existsSync(tarFile)) {
+        throw new Error(`${tarFile} does not exist`)
+      }
+
+      cp.execSync(`tar -xf ${tarFile}.test -C ${tmpdir}`);
+      fs.copyFileSync(path.join(tmpdir, binaryName), path.join(bin, binaryName));
+    } catch (error) {
+      console.error('failed to extract binary.', error.message)
+
+      retries += 1;
+      if (retries <= MAX_RETRIES) {
+        install();
+      }
+    }
   });
 };
